chore: Fix prettier errors

+++ pkg/pool-linear/contracts/ola/OlaLinearPool.sol
@@ -17,11 +17,12 @@ pragma experimental ABIEncoderV2;
 
 import "../LinearPool.sol";
 
-
 interface IOTokenForOlaLinearPool {
     function underlying() external view returns (address);
-    function decimals() external view returns (uint);
-    function exchangeRateStored() external view returns (uint);
+
+    function decimals() external view returns (uint256);
+
+    function exchangeRateStored() external view returns (uint256);
 }
 
 /**
@@ -29,7 +30,7 @@ interface IOTokenForOlaLinearPool {
  * Based on : https://etherscan.io/address/0x2BBf681cC4eb09218BEe85EA2a5d3D13Fa40fC0C#code
  */
 contract OlaLinearPool is LinearPool {
-    uint private immutable _exchangeRateScale;
+    uint256 private immutable _exchangeRateScale;
     IOTokenForOlaLinearPool private _oToken;
 
     constructor(
@@ -44,32 +45,32 @@ contract OlaLinearPool is LinearPool {
         uint256 bufferPeriodDuration,
         address owner
     )
-    LinearPool(
-        vault,
-        name,
-        symbol,
-        mainToken,
-        wrappedToken,
-        upperTarget,
-        swapFeePercentage,
-        pauseWindowDuration,
-        bufferPeriodDuration,
-        owner
-    )
+        LinearPool(
+            vault,
+            name,
+            symbol,
+            mainToken,
+            wrappedToken,
+            upperTarget,
+            swapFeePercentage,
+            pauseWindowDuration,
+            bufferPeriodDuration,
+            owner
+        )
     {
         IOTokenForOlaLinearPool oToken = IOTokenForOlaLinearPool(address(wrappedToken));
         // Sanity check -- Underlying must be the 'mainToken'
         address underlying = oToken.underlying();
         _require(underlying == address(mainToken), Errors.TOKENS_MISMATCH);
 
-        uint oTokenDecimals = oToken.decimals();
-        uint underlyingDecimals = ERC20(underlying).decimals();
+        uint256 oTokenDecimals = oToken.decimals();
+        uint256 underlyingDecimals = ERC20(underlying).decimals();
         require((underlyingDecimals <= 18) && (oTokenDecimals <= 18), "Unsupported decimals");
 
         /**
          * @dev Standard oToken contracts: _exchangeRateScale = 10 + underlyingDecimals;
-         *   oTokens generally have 8 decimals, so this exchange rate takes into consideration the 
-         *   scaling from underlying to oToken. (i.e. 18 decimals -> 8 decimals) This also accounts for 
+         *   oTokens generally have 8 decimals, so this exchange rate takes into consideration the
+         *   scaling from underlying to oToken. (i.e. 18 decimals -> 8 decimals) This also accounts for
          *   oToken markets which may have different than 8 decimals in the future (unlikely).
          */
         _exchangeRateScale = 18 + underlyingDecimals - oTokenDecimals;
@@ -77,11 +78,11 @@ contract OlaLinearPool is LinearPool {
     }
 
     function _getWrappedTokenRate() internal view override returns (uint256) {
-        uint wantedScale = 18;
+        uint256 wantedScale = 18;
 
         // This function returns a 18 decimal fixed point number, but `rate` has [underlying.decimals + 10] decimals
         // so we need to convert it.
-        uint exchangeRateStored = _oToken.exchangeRateStored();
-        return (exchangeRateStored * (10 ** wantedScale)) / (10 ** _exchangeRateScale);
+        uint256 exchangeRateStored = _oToken.exchangeRateStored();
+        return (exchangeRateStored * (10**wantedScale)) / (10**_exchangeRateScale);
     }
 }
